Formalised About section instead of vague Mainbody
